requirements.txt
certifi==2018.10.15
chardet==3.0.4
Django==1.11.16
django-rest-framework==0.1.0
djangorestframework==3.9.0
idna==2.7
pytz==2018.7
requests==2.20.0
urllib3==1.24.1
